ContactsTest.php の テストの追加 $response->assertStatus(403); の追加に伴う修正 #13

+++ app/Http/Controllers/ContactsController.php
@@ -15,14 +15,12 @@ class ContactsController extends Controller
 
     public function store()
     {
-        Contact::create($this->validateData());
+        request()->user()->contacts()->create($this->validateData());
     }
 
     public function show(Contact $contact)
     {
-        // dd($contact->user);
         if (request()->user()->isNot($contact->user)) {
-            //dd('here');
             return response([], 403);
         }
         return $contact;
@@ -30,11 +28,17 @@ class ContactsController extends Controller
 
     public function update(Contact $contact)
     {
+        if (request()->user()->isNot($contact->user)) {
+            return response([], 403);
+        }
         $contact->update($this->validateData());
     }
 
     public function destroy(Contact $contact)
     {
+        if (request()->user()->isNot($contact->user)) {
+            return response([], 403);
+        }
         $contact->delete();
     }
